pyproject.toml
[tool.poetry]
name = "dajare_detector"
version = "1.0.0"
description = "dajare detector"
authors = ["vaaaaanquish <6syun9@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/vaaaaanquish/dajare-detector"
repository = "https://github.com/vaaaaanquish/dajare-detector"

[tool.poetry.dependencies]
python = "^3.7.0"
gokart = "=0.3.19"
swifter = "=1.0.7"
pandas = "=1.1.2"
scikit-learn = "=0.23.2"
luigi = "=3.0.2"
optuna = "=2.3.0"
lightgbm = "=3.1.0"
fuzzysearch = "=0.7.3"
pykakasi = "=2.0.1"
numpy = "=1.19.2"
requests = "=2.24.0"
"alkana.py" = "=0.0.1"
sentence-transformers = { git="https://github.com/sonoisa/sentence-transformers.git" }
neologdn = "=0.4"
tqdm = "=4.50.0"
mecab = "=0.996.2"
mecab-python3 = "=0.996.2"
regex = "=2020.10.28"

[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"
